New Controller handler, Player Class Documenting
TODO: Add back Wilderness & Godwars support

A much more modern approach to Controller handling.

The Player class now has all documentation to further understand Objects, Methods, etc... Also removed some redundant Objects related to NPC & Player.

+++ #633 - Server/combat/com/rs/game/npc/global/GenericNPCSignature.java
@@ -11,7 +11,7 @@ public @interface GenericNPCSignature {
 
 	public int[] npcId();
 	
-	public int mapAreaNameHash() default -1;
+	public byte mapAreaNameHash() default -1;
 	
 	public boolean canBeAttackFromOutOfArea();
